Display active booked appointments slightly different, not perfect yet. Added permalink for ical, via userId since we dont have anything else yet.
