+++ client/styles/styles/styles.less
@@ -58,6 +58,11 @@
   background-color: @gray-lightest;
 }
 
+.fc-list-item.booked.active {
+  font-style: inherit;
+  background-color: @gray-lightest;
+}
+
 .fc-state-default {
   background-image: none;
   text-shadow: none;
