+++ client/views/home_private/appointments/appointments.html
@@ -21,11 +21,13 @@
                 <h3 class="panel-title">Appointments</h3>
             </div>
             <div id="dataview-controls" class="form-group">
-                <button type="submit" id="dataview-download-button" class="btn btn-primary">
+                <form action="/ical/{{userId}}">
+                    <button type="submit" id="dataview-download-button" class="btn btn-primary">
                 <span class="fa fa-download" aria-hidden="true">
                 </span>
-                    Download as Ics
-                </button>
+                        Download as Ics
+                    </button>
+                </form>
             </div>
             {{> ReactiveFullcalendar options=appointmentsCalendarOptions }}
         </div>
