+++ lib/global_utils.js
@@ -58,6 +58,9 @@ this.getCalendarEvents = function (events_in, calendars_in, private_in) {
             if (availability.bookedByConfirmed) {
                 status = "booked";
                 color = "#FF0000";
+                if (moment(availability.startDate)>moment()){
+                    status = status + " active";
+                }
             } else if (moment(availability.bookedByDate) >= moment().add(-reservationThreshold, 'm') && !availability.bookedByConfirmed && availability.bookedByDate) {
                 status = "reserved";
                 color = "#FFFF00";
@@ -83,7 +86,7 @@ this.getCalendarEvents = function (events_in, calendars_in, private_in) {
                     title = "unknown calendar";
                     color = "000"
                 }
-                if (status === "booked" || status === "reserved") {
+                if (status === "booked" || status === "booked active" || status === "reserved") {
                     if (calendar = calendars_in.findOne({_id: availability.bookedByCalendarId.toString()})) {
                         title = calendar.name;
                         color = calendar.color;
@@ -115,7 +118,7 @@ this.getCalendarEvents = function (events_in, calendars_in, private_in) {
 
 this.calendarClickOptions = function (calEvent, pageSession) {
     if (calEvent.private) {
-        if (calEvent.status === "booked" || calEvent.status === "reserved") {
+        if (calEvent.status === "booked" || calEvent.status === "reserved" || calEvent.status === "booked active") {
             Router.go("home_private.appointment", {_appointmentId: calEvent.id});
         } else if (calEvent.status === "available" && calEvent.start >= new Date()) {
             Router.go("home_private.edit_availability", {_availabilityId: calEvent.id});
