+++ server/controllers/router.js
@@ -1,3 +1,41 @@
-Router.map(function () {
+import {Availabilities} from '/imports/api/availabilitiesCollection';
+import {Calendars} from '/imports/api/calendarsCollection';
 
+Router.map(function () {
+    Router.route('ical', {
+        name: 'ical',
+        path: "ical/:_token",
+        where: 'server',
+        waitOn: function () {
+            return [
+                Meteor.subscribe('allAppointments'),
+                Meteor.subscribe('allCalendars')
+            ]
+        },
+        action: function() {
+            var user = Meteor.users.findOne(this.params._token);
+            var currentEvents = getCalendarEvents(Availabilities.find({userId: user._id,startDate: {$gt: new Date()},bookedByConfirmed: true}).fetch(), Calendars, true);
+            var calendar = new IcsGenerator({prodId: "//MeteorKalender//Frankfurt University of Applied Sciences",
+                method: "REQUEST",
+            });
+            for (var i = 0;i<currentEvents.length;i++){
+                var newEvent = calendar.createEvent({
+                    uid: currentEvents[i].id+"@meteorkalender",
+                    summary: currentEvents[i].title,
+                    dtStart: currentEvents[i].start,
+                    dtEnd: currentEvents[i].end,
+                    organizer: {cn: user.profile.name, mailTo: user.emails[0].address},
+                    attendees: [
+                       {cn: currentEvents[i].attendee_name, mailTo: currentEvents[i].attendee_mail}
+                    ]
+                });
+                calendar.addEvent(newEvent);
+            }
+            this.response.writeHead(200, {
+                'Content-Type': 'text/plain;charset=utf-8'
+            });
+            this.response.write(calendar.toIcsString());
+            this.response.end();
+        }
+    });
 });
